pyproject.toml
[tool.poetry]
name = "molgen"
version = "1.0a1"
description = "Auto-regressive molecule generator"
authors = ["Stephan Holzgruber <stephan.holzgruber@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
codetiming = "^1.4.0"
# fcd-torch = {version = "^1.0.7", platform = "darwin"}
# fcd = {version = "^1.1", platform = "linux"}
fcd-torch = "^1.0.7"
humanfriendly = "^10.0"
hydra-core = "^1.2"
jupyter = "^1.0.0"
loguru = "^0.6.0"
notebook = "^6.4.11"
numpy = "^1.22.3"
packaging = "^21.3"
pandas = "^1.4.3"
python = ">=3.9,<3.10"
PyTDC = "^0.3.6"
pytorch-lightning = "^1.7.0"
rdkit-pypi = "^2022.3.2"
rich = "^12.4.4"
scipy = "^1.8.0"
subset = "^0.1.2"
# tensorflow = "2.9.2"
# torch = "^1.12"
# torch = "^1.13.0"
torch = [
  { version = "1.12.0", platform = "darwin" },
  { version = "^1.13.0", platform = "linux" }
]
torchmetrics = "^0.9.2"
# torchtext = "^0.13.0"
tqdm = "^4.64.0"
wandb = "^0.13.0"
transformers = "^4.24.0"
datasets = "^2.6.1"
evaluate = "^0.3.0"
tensorboard = "^2.11.0"
rdchiral = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0"
isort = "^5.10.1"
mypy = "^0.991"
psutil = "^5.9.4"
pytest = "^7.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

# see https://towardsdev.com/how-to-manage-pytorch-dependency-with-poetry-14da0057af5a
# [tool.poe.tasks]
# install-pytorch = "pip install torch==1.13 -f https://download.pytorch.org/whl/torch_stable.html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
)/
'''

[tool.isort]
profile = "black"
line_length = 80
